+++ CHANGES.rst
@@ -5,6 +5,10 @@ Moved to github. So a few files are renamed from .txt to .rst
 
 Added .xhtml_to_html() and .remove_namespaces()
 
+Use requests to fetch urls (if available)
+
+Use restkit's proxy instead of Paste (which will die with py3)
+
 python2.5 is no longer supported (may work, but tests are broken)
 
 1.2.3